# stop: grow stacks for async tasks