@tbg 2024-12-19
Before this PR

<img width="971" alt="image" src="https://github.com/user-attachments/assets/c830ef45-0ed4-4cc2-9493-c67fb17ec69d" />

This was running on an older SHA, so the comparison isn't scientific, but I anecdotally see around 2-3% even on recent SHAs.

With this PR:

-TODO-