@tbg 2024-12-23
I ran `oltp_read_write` 5x on https://github.com/cockroachdb/cockroach/pull/137927, but it's so choppy:

```
base: 21778 22081 20964 19356 17538
with: 20764 20481 18452 20808 21295
```

hoping that https://github.com/cockroachdb/cockroach/pull/137930 significantly reduces the choppiness, and will revisit in the new year.