@tbg 2025-01-14
10x sysbench-settings/oltp_read_write on this branch:

```
$ rg 'queries:' | grep -Eo '[0-9]+\.[0-9]+ per sec'
22806.27 per sec
23212.83 per sec
23955.21 per sec
22117.93 per sec
22592.12 per sec
23299.67 per sec
23590.72 per sec
23312.14 per sec
23451.91 per sec
23201.54 per sec
```

Now running it 10x at the preceding commit.