@RaduBerinde 2025-01-17
Does `Grow()` ensure 32KB of stack? Isn't that still pretty low?