@tbg 2025-01-17
> Does Grow() ensure 32KB of stack? Isn't that still pretty low?

It places a 32kb object on the stack, which should grow the stack to 64kb. It's pretty small still, but seems to provide a large benefit. I'm hesitant to make stacks too large by default, since that carries a cost too.
I think Nathan had some analysis on some pebble issue about the actual stack depths we could expect, but experimentally, this change seems to help a lot.

bors r+