@RaduBerinde 2025-01-17
I wonder if it would be worth trying to mess with Go's policy for stack growth (instead of doubling each time, we could have fewer thresholds, e.g. 32K, 128KB, 512KB) in our Go patch.