@tbg 2025-01-17
It's a good question - I'm not sure.

@nvanbenschoten had this great analysis of stack offsets here: https://github.com/cockroachdb/pebble/pull/4187#issuecomment-2514997541

For CRDB he told me this is how he analyzed it:

```
go tool objdump -S cockroach > cockroach.s
rg -B1 'SUB R\d\d, RSP' cockroach.s | rg '\$' | awk '{print $4}' | sed 's/[$,]//g' | sort -nr | head -n200 > large.txt
```

but I think the approach in pebble should work too:

```
go build -gcflags -S ./pkg/cmd/cockroach
```

and then proceed as in the issue. (Of course this wouldn't be our official bazel build, though if `go` is our patched go it might come pretty close for most intents and purposes, but likely no longer once pgo comes into play).

I don't know if Nathan posted the analysis results for cockroach anywhere, but I knew he ran it, so at least there wasn't anything obvious. Knowing the stack size for each function, we could, for, say, a CPU profile, compute the actual stack depths for each stack as the sum of that of their stack frames.